refactored WebTestCase into a trait
