+++ Test/WebTestCase.php
@@ -17,30 +17,5 @@ use Symfony\Bundle\FrameworkBundle\Test\WebTestCase as BaseWebTestCase;
  */
 abstract class WebTestCase extends BaseWebTestCase
 {
-    protected function assertJsonResponse($response, $statusCode = 200)
-    {
-        $this->assertEquals(
-            $statusCode, $response->getStatusCode(),
-            $response->getContent()
-        );
-        $this->assertTrue(
-            $response->headers->contains('Content-Type', 'application/json'),
-            $response->headers
-        );
-    }
-
-    protected function jsonRequest($verb, $endpoint, array $data = array())
-    {
-        $data = empty($data) ? null : json_encode($data);
-
-        return $this->client->request($verb, $endpoint,
-            array(),
-            array(),
-            array(
-                'HTTP_ACCEPT'  => 'application/json',
-                'CONTENT_TYPE' => 'application/json'
-            ),
-            $data
-        );
-    }
+    use JsonTestTrait;
 }
